Error handling adds functionality for PSU, abstract fans, display, fan power enable/disable, and more

+++ firmware/lib/core/include/core_defines.h
@@ -9,7 +9,7 @@
 
 namespace CogCore {
 
-typedef int32_t TimeMs;
+typedef uint64_t TimeMs;
 typedef int32_t TaskId;
 
 enum class TaskState {
